Code Review - P3
Good job on your project! All the functions are clear and easy to find, and I like that you neatly separate file in different folders so it's clearly organized. The search functionality and UI looked great as well. 

As a small advice, it's always a good idea to protect your private pages from unauthorized access. For example even without logging in I can go to https://vital-stqk.onrender.com/portal and start booking appointments. 

Secondly for some reason the social media links were not working in the footer for. Might want to check on that :)